requirements.txt
pandas
tabulate
fastapi
uvicorn[standard]
langchain
langchain-google-genai
python-dotenv